requirements.txt
flask
python-dotenv
requests
google-generativeai
newspaper3k
gunicorn==21.x.x